WIp
Current approach has to be abandoned.
Approach: the plugin functionality is triggered from the model. (model has the previewable description)
Problem Observed: In any (web) request, controller is loaded before the models. So some triggers to the functionality have to go in it.
Concept: Trying to access controller from model is againsts MVC